CIX-Added-Rating-Screen-Design-Api

+++ app/src/main/java/com/harshilpadsala/watchlistx/data/req/MediaRatingRequest.kt
@@ -1,5 +1,5 @@
 package com.harshilpadsala.watchlistx.data.req
 
 data class MediaRatingRequest(
-    val value : Double
+    val value : Int
 )
